Arreglo para evitar codigos de materias negativos

+++ flowa/database/migrations/2024_01_16_235254_create_materias_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
 
             $table->string('nombre_materia')->unique();
-            $table->bigInteger('codigo')->unique();
+            $table->unsignedInteger('codigo')->unique();
             
             $table->unsignedBigInteger('profesor_id');        
             $table->foreign('profesor_id')->references('id')->on('profesors')->onDelete('cascade');
